Added Lorempixel check for ImageTest.php

+++ test/Faker/Provider/ImageTest.php
@@ -45,18 +45,30 @@ class ImageTest extends \PHPUnit_Framework_TestCase
 
     public function testDownloadWithDefaults()
     {
-        $file = Image::image(sys_get_temp_dir());
-        $this->assertFileExists($file);
-        if (function_exists('getimagesize')) {
-            list($width, $height, $type, $attr) = getimagesize($file);
-            $this->assertEquals(640, $width);
-            $this->assertEquals(480, $height);
-            $this->assertEquals(constant('IMAGETYPE_JPEG'), $type);
-        } else {
-            $this->assertEquals('jpg', pathinfo($file, PATHINFO_EXTENSION));
-        }
-        if (file_exists($file)) {
-            unlink($file);
+        $url = "http://www.lorempixel.com/";
+        $curlPing = curl_init($url);
+        curl_setopt($curlPing, CURLOPT_TIMEOUT, 5);
+        curl_setopt($curlPing, CURLOPT_CONNECTTIMEOUT, 5);
+        curl_setopt($curlPing, CURLOPT_RETURNTRANSFER, true);
+        $data = curl_exec($curlPing);
+        $httpCode = curl_getinfo($curlPing, CURLINFO_HTTP_CODE);
+        curl_close($curlPing);
+
+        if ($httpCode >= 200 && $httpCode < 300)
+        {
+            $file = Image::image(sys_get_temp_dir());
+            $this->assertFileExists($file);
+            if (function_exists('getimagesize')) {
+                list($width, $height, $type, $attr) = getimagesize($file);
+                $this->assertEquals(640, $width);
+                $this->assertEquals(480, $height);
+                $this->assertEquals(constant('IMAGETYPE_JPEG'), $type);
+            } else {
+                $this->assertEquals('jpg', pathinfo($file, PATHINFO_EXTENSION));
+            }
+            if (file_exists($file)) {
+                unlink($file);
+            }
         }
     }
 }
